endpoints ok

+++ src/modules/delete_institute_by_id/app/delete_institute_by_id_controller.ts
@@ -1,32 +1,72 @@
-import { IRequest } from "src/shared/helpers/external_interfaces/external_interface";
+import { IRequest } from 'src/shared/helpers/external_interfaces/external_interface';
 import {
+  BadRequest,
+  Conflict,
   InternalServerError,
   NotFound,
   OK,
-} from "src/shared/helpers/external_interfaces/http_codes";
-import { NoItemsFound } from "src/shared/helpers/errors/errors";
-import { DeleteInstituteByIdViewModel } from "./delete_institute_by_id_viewmodel";
-import { DeleteInstituteByIdUseCase } from "./delete_institute_by_id_usecase";
+  Unauthorized,
+} from 'src/shared/helpers/external_interfaces/http_codes';
+import {
+  ConflictItems,
+  EntityError,
+  ForbiddenAction,
+  MissingParameters,
+  NoItemsFound,
+  WrongTypeParameters,
+} from 'src/shared/helpers/errors/errors';
+import { DeleteInstituteByIdViewModel } from './delete_institute_by_id_viewmodel';
+import { DeleteInstituteByIdUseCase } from './delete_institute_by_id_usecase';
+import { UserAPIGatewayDTO } from 'src/shared/infra/database/dtos/user_api_gateway_dto';
+import { ROLE_TYPE } from 'src/shared/domain/enums/role_type_enum';
 
 export class DeleteInstituteByIdController {
   constructor(private readonly usecase: DeleteInstituteByIdUseCase) {}
 
-  async handle(req: IRequest): Promise<any> {
+  async handle(
+    req: IRequest,
+    requesterUser: Record<string, any>
+  ): Promise<any> {
     try {
-      const { instituteId } = req.data;
+      const userApiGateway = UserAPIGatewayDTO.fromAPIGateway(requesterUser);
+
+      if (!userApiGateway) throw new ForbiddenAction('Usuário');
+
+      if (userApiGateway.role !== ROLE_TYPE.OWNER) {
+        throw new ForbiddenAction(
+          'Usuário nao tem permissão para excluir um instituto'
+        );
+      }
+
+      const { instituteId } = req.data.body;
+
       await this.usecase.execute(instituteId as string);
+
       const viewmodel = new DeleteInstituteByIdViewModel(
-        "Instituto deletado com sucesso"
+        'Instituto deletado com sucesso'
       );
       return new OK(viewmodel.toJSON());
     } catch (error: any) {
+      if (
+        error instanceof EntityError ||
+        error instanceof MissingParameters ||
+        error instanceof WrongTypeParameters
+      ) {
+        return new BadRequest(error.message);
+      }
+      if (error instanceof ConflictItems) {
+        return new Conflict(error.message);
+      }
+      if (error instanceof ForbiddenAction) {
+        return new Unauthorized(error.message);
+      }
+
       if (error instanceof NoItemsFound) {
         return new NotFound(error.message);
       }
+
       if (error instanceof Error) {
-        return new InternalServerError(
-          `DeleteInstituteByIdController, Error on handle: ${error.message}`
-        );
+        return new InternalServerError(error.message);
       }
     } finally {
       await this.usecase.repository.closeSession();
